backend-ai/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Services
openai==1.12.0
google-generativeai==0.3.2
azure-cognitiveservices-speech==1.35.0
azure-storage-blob==12.19.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy>=1.26.0

# ML/NLP
transformers>=4.37.2
torch>=2.2.0
nltk==3.8.1
spacy==3.7.4

# Background/Async
celery==5.3.6
redis==5.0.1
aiocache==0.12.2
diskcache==5.6.3
aiohttp==3.9.3
httpx==0.26.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25

# Monitoring
prometheus-client==0.19.0
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
tenacity==8.2.3
